Moved package.json search into library PackageJsonHelpers
